Extract a common implementation of `TaskDependencyResolveContext` that is shared in a couple of places.
